Rename internal/pkg/levee package name (should be 'levee', is 'internal').
